requirements.txt
streamlit
pymupdf
sentence_transformers
faiss-cpu
python-dotenv
ibm-watsonx-ai
fitz
